Move and rename NamingProxy.java
