+++ client/src/main/java/com/alibaba/nacos/client/naming/beat/BeatReactor.java
@@ -23,7 +23,7 @@ import com.alibaba.nacos.api.naming.NamingResponseCode;
 import com.alibaba.nacos.api.naming.pojo.Instance;
 import com.alibaba.nacos.api.naming.utils.NamingUtils;
 import com.alibaba.nacos.client.monitor.MetricsMonitor;
-import com.alibaba.nacos.client.naming.net.NamingProxy;
+import com.alibaba.nacos.client.naming.net.http.NamingHttpClientProxy;
 import com.alibaba.nacos.client.naming.utils.UtilAndComs;
 import com.alibaba.nacos.common.lifecycle.Closeable;
 import com.alibaba.nacos.common.utils.JacksonUtils;
@@ -48,17 +48,17 @@ public class BeatReactor implements Closeable {
     
     private final ScheduledExecutorService executorService;
     
-    private final NamingProxy serverProxy;
+    private final NamingHttpClientProxy serverProxy;
     
     private boolean lightBeatEnabled = false;
     
     public final Map<String, BeatInfo> dom2Beat = new ConcurrentHashMap<String, BeatInfo>();
     
-    public BeatReactor(NamingProxy serverProxy) {
+    public BeatReactor(NamingHttpClientProxy serverProxy) {
         this(serverProxy, UtilAndComs.DEFAULT_CLIENT_BEAT_THREAD_COUNT);
     }
     
-    public BeatReactor(NamingProxy serverProxy, int threadCount) {
+    public BeatReactor(NamingHttpClientProxy serverProxy, int threadCount) {
         this.serverProxy = serverProxy;
         this.executorService = new ScheduledThreadPoolExecutor(threadCount, new ThreadFactory() {
             @Override
