+++ client/src/test/java/com/alibaba/nacos/client/BeatReactorTest.java
@@ -19,7 +19,7 @@ package com.alibaba.nacos.client;
 import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.client.naming.beat.BeatInfo;
 import com.alibaba.nacos.client.naming.beat.BeatReactor;
-import com.alibaba.nacos.client.naming.net.NamingProxy;
+import com.alibaba.nacos.client.naming.net.http.NamingHttpClientProxy;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -34,7 +34,7 @@ import java.util.concurrent.ScheduledThreadPoolExecutor;
 public class BeatReactorTest {
     
     @Mock
-    private NamingProxy namingProxy;
+    private NamingHttpClientProxy namingHttpClientProxy;
     
     @Test
     public void test() throws NoSuchFieldException, IllegalAccessException, InterruptedException, NacosException {
@@ -48,7 +48,7 @@ public class BeatReactorTest {
         beatInfo.setScheduled(false);
         beatInfo.setPeriod(1000L);
     
-        BeatReactor beatReactor = new BeatReactor(namingProxy);
+        BeatReactor beatReactor = new BeatReactor(namingHttpClientProxy);
         beatReactor.addBeatInfo("testService", beatInfo);
         
         Assert.assertEquals(1, getActiveThread(beatReactor));
