+++ test/src/test/java/com/alibaba/nacos/test/naming/NamingHttpClientProxy_ITCase.java
@@ -46,7 +46,7 @@ import java.util.concurrent.TimeUnit;
 @RunWith(SpringRunner.class)
 @SpringBootTest(classes = NamingApp.class, properties = {"server.servlet.context-path=/nacos"},
     webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class NamingProxy_ITCase {
+public class NamingHttpClientProxy_ITCase {
     @LocalServerPort
     private int port;
     @Autowired
